TMDM-6308: Move suspicious error message to DEBUG level (alternative to initial fix) // https://jira.talendforge.org/browse/TMDM-6308
git-svn-id: http://talendforge.org/svn/tom/trunk@111513 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.mdm.core/src/com/amalto/core/jobox/util/JoboxUtil.java
@@ -45,10 +45,14 @@ public class JoboxUtil {
         try {
             delAllFile(folderPath);
             File myFilePath = new File(folderPath);
-            if (!myFilePath.delete()) {
+            if (!myFilePath.exists()) {
                 if (LOGGER.isDebugEnabled()) {
-                    LOGGER.debug("Delete folder failed for '" + folderPath + "'."); //$NON-NLS-1$ //$NON-NLS-2$
+                    LOGGER.debug("Delete folder failed for '" + folderPath + "' (path does not exist)."); //$NON-NLS-1$ //$NON-NLS-2$
                 }
+                return;
+            }
+            if (!myFilePath.delete()) {
+                LOGGER.error("Delete folder failed for '" + folderPath + "'."); //$NON-NLS-1$ //$NON-NLS-2$
             }
         } catch (Exception e) {
             throw new JoboxException(e);
